// ts/styl.types.ts
export type HLBaseStyleType =
	| "bold"
	// | "dim"
	| "italic"
	| "underline"
	| "inverse"
	| "blink"
	| "hidden"
	| "strikethrough";
export type HLStyleType =
	| "bold"
	| "b"
	// | "dim"
	// | "d"
	| "italic"
	| "i"
	| "underline"
	| "u"
	| "inverse"
	| "blink"
	| "hidden"
	| "strikethrough"
	| "s";

export const HLStyleKeys: string[] = [
	"bold",
	"b",
	// "dim",
	// "d",
	"italic",
	"i",
	"underline",
	"u",
	"inverse",
	"blink",
	"hidden",
	"strikethrough",
	"s",
];

export type HLColorType =
	| "white"
	| "grey"
	| "black"
	| "blue"
	| "cyan"
	| "green"
	| "magenta"
	| "red"
	| "yellow";

export const HLColorKeys: string[] = [
	"white",
	"grey",
	"black",
	"blue",
	"cyan",
	"green",
	"magenta",
	"red",
	"yellow",
];

export type HLBackColorType =
	| "whiteBG"
	| "greyBG"
	| "blackBG"
	| "blueBG"
	| "cyanBG"
	| "greenBG"
	| "magentaBG"
	| "redBG"
	| "yellowBG";

export const HLBackColorKeys: string[] = [
	"whiteBG",
	"greyBG",
	"blackBG",
	"blueBG",
	"cyanBG",
	"greenBG",
	"magentaBG",
	"redBG",
	"yellowBG",
];

export type HLColorsType =
	| "white"
	| "grey"
	| "black"
	| "blue"
	| "cyan"
	| "green"
	| "magenta"
	| "red"
	| "yellow"
	| "whiteBG"
	| "greyBG"
	| "blackBG"
	| "blueBG"
	| "cyanBG"
	| "greenBG"
	| "magentaBG"
	| "redBG"
	| "yellowBG";

export const HLColorsKeys: string[] = [
	"white",
	"grey",
	"black",
	"blue",
	"cyan",
	"green",
	"magenta",
	"red",
	"yellow",
	"whiteBG",
	"greyBG",
	"blackBG",
	"blueBG",
	"cyanBG",
	"greenBG",
	"magentaBG",
	"redBG",
	"yellowBG",
];